Create template for sensory data

Create template for inputting data that collected from sensory research.

create_template(data, attribute)

Arguments

data
output from prepare function

attribute
a numeric value of number or sensory attributes or vector of sensory terms (lexicon)

Value

Returns a dataframe of sensory design with panelist, presentation order, product, blind code (optional) and sensory attributes columns

Examples

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # Design of Experiment: <20 x 6>
#> # Panelist:             20 subjects
#> # Product:              5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_03 (641) prod_01 (533) prod_02 (495) prod_05 (272) prod_04 (79…
#>  2 ind_002  prod_01 (533) prod_05 (272) prod_03 (641) prod_04 (797) prod_02 (49…
#>  3 ind_003  prod_05 (272) prod_04 (797) prod_01 (533) prod_02 (495) prod_03 (64…
#>  4 ind_004  prod_04 (797) prod_02 (495) prod_05 (272) prod_03 (641) prod_01 (53…
#>  5 ind_005  prod_02 (495) prod_03 (641) prod_04 (797) prod_01 (533) prod_05 (27…
#>  6 ind_006  prod_04 (797) prod_05 (272) prod_02 (495) prod_01 (533) prod_03 (64…
#>  7 ind_007  prod_02 (495) prod_04 (797) prod_03 (641) prod_05 (272) prod_01 (53…
#>  8 ind_008  prod_03 (641) prod_02 (495) prod_01 (533) prod_04 (797) prod_05 (27…
#>  9 ind_009  prod_01 (533) prod_03 (641) prod_05 (272) prod_02 (495) prod_04 (79…
#> 10 ind_010  prod_05 (272) prod_01 (533) prod_04 (797) prod_03 (641) prod_02 (49…
#> 11 ind_011  prod_03 (641) prod_01 (533) prod_02 (495) prod_05 (272) prod_04 (79…
#> 12 ind_012  prod_01 (533) prod_05 (272) prod_03 (641) prod_04 (797) prod_02 (49…
#> 13 ind_013  prod_05 (272) prod_04 (797) prod_01 (533) prod_02 (495) prod_03 (64…
#> 14 ind_014  prod_04 (797) prod_02 (495) prod_05 (272) prod_03 (641) prod_01 (53…
#> 15 ind_015  prod_02 (495) prod_03 (641) prod_04 (797) prod_01 (533) prod_05 (27…
#> 16 ind_016  prod_04 (797) prod_05 (272) prod_02 (495) prod_01 (533) prod_03 (64…
#> 17 ind_017  prod_02 (495) prod_04 (797) prod_03 (641) prod_05 (272) prod_01 (53…
#> 18 ind_018  prod_03 (641) prod_02 (495) prod_01 (533) prod_04 (797) prod_05 (27…
#> 19 ind_019  prod_01 (533) prod_03 (641) prod_05 (272) prod_02 (495) prod_04 (79…
#> 20 ind_020  prod_05 (272) prod_01 (533) prod_04 (797) prod_03 (641) prod_02 (49…

create_template(design, attribute = c("Sweetness", "Mint", "Green", "Herbal"))

#> # A sensory table: <100 x 8>
#> # Panelist:        panelist <20 subjects>
#> # Product:         product <5 items>
#> # Attribute:       Sweetness, Mint, Green, Herbal<4 lexicons>
#>    panelist pres_order product blind_code Sweetness Mint  Green Herbal
#>    <chr>    <chr>      <chr>   <chr>      <chr>     <chr> <chr> <chr> 
#>  1 ind_001  order_1    prod_03 641        NA        NA    NA    NA    
#>  2 ind_001  order_2    prod_01 533        NA        NA    NA    NA    
#>  3 ind_001  order_3    prod_02 495        NA        NA    NA    NA    
#>  4 ind_001  order_4    prod_05 272        NA        NA    NA    NA    
#>  5 ind_001  order_5    prod_04 797        NA        NA    NA    NA    
#>  6 ind_002  order_1    prod_01 533        NA        NA    NA    NA    
#>  7 ind_002  order_2    prod_05 272        NA        NA    NA    NA    
#>  8 ind_002  order_3    prod_03 641        NA        NA    NA    NA    
#>  9 ind_002  order_4    prod_04 797        NA        NA    NA    NA    
#> 10 ind_002  order_5    prod_02 495        NA        NA    NA    NA    
#> # … with 90 more rows