Create template for data input

Create template for inputting data collected from sensory research

create_template_(.data, panelist, attribute)

Arguments

.data
output from generate_design function

panelist
column containing panelist information

attribute
a numeric value of number or sensory attributes or vector of sensory terms (lexicon)

Value

Returns a dataframe of sensory design with Panelist, Order, Product, and Sensory Attributes columns

Examples

design <- generate_design_(n_panelist = 20, product = 5, use_coding = TRUE)
design

#> # A tibble: 20 x 6
#>    Panelis   Urutan1       Urutan2       Urutan3       Urutan4      Urutan5     
#>    <chr>     <chr>         <chr>         <chr>         <chr>        <chr>       
#>  1 Panelis1  Produk1 (474) Produk2 (724) Produk5 (211) Produk3 (23… Produk4 (90…
#>  2 Panelis2  Produk2 (724) Produk3 (233) Produk1 (474) Produk4 (90… Produk5 (21…
#>  3 Panelis3  Produk3 (233) Produk4 (904) Produk2 (724) Produk5 (21… Produk1 (47…
#>  4 Panelis4  Produk4 (904) Produk5 (211) Produk3 (233) Produk1 (47… Produk2 (72…
#>  5 Panelis5  Produk5 (211) Produk1 (474) Produk4 (904) Produk2 (72… Produk3 (23…
#>  6 Panelis6  Produk4 (904) Produk3 (233) Produk5 (211) Produk2 (72… Produk1 (47…
#>  7 Panelis7  Produk5 (211) Produk4 (904) Produk1 (474) Produk3 (23… Produk2 (72…
#>  8 Panelis8  Produk1 (474) Produk5 (211) Produk2 (724) Produk4 (90… Produk3 (23…
#>  9 Panelis9  Produk2 (724) Produk1 (474) Produk3 (233) Produk5 (21… Produk4 (90…
#> 10 Panelis10 Produk3 (233) Produk2 (724) Produk4 (904) Produk1 (47… Produk5 (21…
#> 11 Panelis11 Produk1 (474) Produk2 (724) Produk5 (211) Produk3 (23… Produk4 (90…
#> 12 Panelis12 Produk2 (724) Produk3 (233) Produk1 (474) Produk4 (90… Produk5 (21…
#> 13 Panelis13 Produk3 (233) Produk4 (904) Produk2 (724) Produk5 (21… Produk1 (47…
#> 14 Panelis14 Produk4 (904) Produk5 (211) Produk3 (233) Produk1 (47… Produk2 (72…
#> 15 Panelis15 Produk5 (211) Produk1 (474) Produk4 (904) Produk2 (72… Produk3 (23…
#> 16 Panelis16 Produk4 (904) Produk3 (233) Produk5 (211) Produk2 (72… Produk1 (47…
#> 17 Panelis17 Produk5 (211) Produk4 (904) Produk1 (474) Produk3 (23… Produk2 (72…
#> 18 Panelis18 Produk1 (474) Produk5 (211) Produk2 (724) Produk4 (90… Produk3 (23…
#> 19 Panelis19 Produk2 (724) Produk1 (474) Produk3 (233) Produk5 (21… Produk4 (90…
#> 20 Panelis20 Produk3 (233) Produk2 (724) Produk4 (904) Produk1 (47… Produk5 (21…

create_template_(design, panelist = Panelis, attribute = c("Sweetness", "Mint", "Green", "Herbal"))

#> # A tibble: 100 x 7
#>    Panelis  Urutan  Produk        Sweetness  Mint Green Herbal
#>    <fct>    <chr>   <chr>             <int> <int> <int>  <int>
#>  1 Panelis1 Urutan1 Produk1 (474)        NA    NA    NA     NA
#>  2 Panelis1 Urutan2 Produk2 (724)        NA    NA    NA     NA
#>  3 Panelis1 Urutan3 Produk5 (211)        NA    NA    NA     NA
#>  4 Panelis1 Urutan4 Produk3 (233)        NA    NA    NA     NA
#>  5 Panelis1 Urutan5 Produk4 (904)        NA    NA    NA     NA
#>  6 Panelis2 Urutan1 Produk2 (724)        NA    NA    NA     NA
#>  7 Panelis2 Urutan2 Produk3 (233)        NA    NA    NA     NA
#>  8 Panelis2 Urutan3 Produk1 (474)        NA    NA    NA     NA
#>  9 Panelis2 Urutan4 Produk4 (904)        NA    NA    NA     NA
#> 10 Panelis2 Urutan5 Produk5 (211)        NA    NA    NA     NA
#> # … with 90 more rows