Generate Design

Generate design of presentation order based on Williams design.

generate_design(n_panelist, product, use_coding = TRUE)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

use_coding
wheteher to generate random three digit number for labeling

Value

Returns a dataframe of sensory design with Panelist column and Order of product presentation columns

Examples

generate_design(n_panelist = 10, product = letters[1:14], use_coding = FALSE)

#> # A tibble: 10 x 15
#>    Panelist Order1 Order2 Order3 Order4 Order5 Order6 Order7 Order8 Order9
#>    <chr>    <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr> 
#>  1 Panelis… a      b      n      c      m      d      l      e      k     
#>  2 Panelis… b      c      a      d      n      e      m      f      l     
#>  3 Panelis… c      d      b      e      a      f      n      g      m     
#>  4 Panelis… d      e      c      f      b      g      a      h      n     
#>  5 Panelis… e      f      d      g      c      h      b      i      a     
#>  6 Panelis… f      g      e      h      d      i      c      j      b     
#>  7 Panelis… g      h      f      i      e      j      d      k      c     
#>  8 Panelis… h      i      g      j      f      k      e      l      d     
#>  9 Panelis… i      j      h      k      g      l      f      m      e     
#> 10 Panelis… j      k      i      l      h      m      g      n      f     
#> # … with 5 more variables: Order10 <chr>, Order11 <chr>, Order12 <chr>,
#> #   Order13 <chr>, Order14 <chr>

generate_design(n_panelist = 10, product = letters[1:14], use_coding = TRUE)

#> # A tibble: 10 x 15
#>    Panelist Order1 Order2 Order3 Order4 Order5 Order6 Order7 Order8 Order9
#>    <chr>    <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr> 
#>  1 Panelis… a (36… b (30… n (98… c (73… m (69… d (56… l (42… e (69… k (63…
#>  2 Panelis… b (30… c (73… a (36… d (56… n (98… e (69… m (69… f (92… l (42…
#>  3 Panelis… c (73… d (56… b (30… e (69… a (36… f (92… n (98… g (35… m (69…
#>  4 Panelis… d (56… e (69… c (73… f (92… b (30… g (35… a (36… h (78… n (98…
#>  5 Panelis… e (69… f (92… d (56… g (35… c (73… h (78… b (30… i (81… a (36…
#>  6 Panelis… f (92… g (35… e (69… h (78… d (56… i (81… c (73… j (32… b (30…
#>  7 Panelis… g (35… h (78… f (92… i (81… e (69… j (32… d (56… k (63… c (73…
#>  8 Panelis… h (78… i (81… g (35… j (32… f (92… k (63… e (69… l (42… d (56…
#>  9 Panelis… i (81… j (32… h (78… k (63… g (35… l (42… f (92… m (69… e (69…
#> 10 Panelis… j (32… k (63… i (81… l (42… h (78… m (69… g (35… n (98… f (92…
#> # … with 5 more variables: Order10 <chr>, Order11 <chr>, Order12 <chr>,
#> #   Order13 <chr>, Order14 <chr>

design <- generate_design(n_panelist = 20, product = 5, use_coding = TRUE)
design

#> # A tibble: 20 x 6
#>    Panelist   Order1        Order2        Order3       Order4       Order5      
#>    <chr>      <chr>         <chr>         <chr>        <chr>        <chr>       
#>  1 Panelist1  Product1 (47… Product2 (72… Product5 (2… Product3 (2… Product4 (9…
#>  2 Panelist2  Product2 (72… Product3 (23… Product1 (4… Product4 (9… Product5 (2…
#>  3 Panelist3  Product3 (23… Product4 (90… Product2 (7… Product5 (2… Product1 (4…
#>  4 Panelist4  Product4 (90… Product5 (21… Product3 (2… Product1 (4… Product2 (7…
#>  5 Panelist5  Product5 (21… Product1 (47… Product4 (9… Product2 (7… Product3 (2…
#>  6 Panelist6  Product4 (90… Product3 (23… Product5 (2… Product2 (7… Product1 (4…
#>  7 Panelist7  Product5 (21… Product4 (90… Product1 (4… Product3 (2… Product2 (7…
#>  8 Panelist8  Product1 (47… Product5 (21… Product2 (7… Product4 (9… Product3 (2…
#>  9 Panelist9  Product2 (72… Product1 (47… Product3 (2… Product5 (2… Product4 (9…
#> 10 Panelist10 Product3 (23… Product2 (72… Product4 (9… Product1 (4… Product5 (2…
#> 11 Panelist11 Product1 (47… Product2 (72… Product5 (2… Product3 (2… Product4 (9…
#> 12 Panelist12 Product2 (72… Product3 (23… Product1 (4… Product4 (9… Product5 (2…
#> 13 Panelist13 Product3 (23… Product4 (90… Product2 (7… Product5 (2… Product1 (4…
#> 14 Panelist14 Product4 (90… Product5 (21… Product3 (2… Product1 (4… Product2 (7…
#> 15 Panelist15 Product5 (21… Product1 (47… Product4 (9… Product2 (7… Product3 (2…
#> 16 Panelist16 Product4 (90… Product3 (23… Product5 (2… Product2 (7… Product1 (4…
#> 17 Panelist17 Product5 (21… Product4 (90… Product1 (4… Product3 (2… Product2 (7…
#> 18 Panelist18 Product1 (47… Product5 (21… Product2 (7… Product4 (9… Product3 (2…
#> 19 Panelist19 Product2 (72… Product1 (47… Product3 (2… Product5 (2… Product4 (9…
#> 20 Panelist20 Product3 (23… Product2 (72… Product4 (9… Product1 (4… Product5 (2…