Prepare sensory experiment

Generate a design of presentation order based on Williams design.

prepare(n_panelist, product, blind_code = FALSE, seed = NULL)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

blind_code
wheteher to generate random three digit number for labeling

seed
an integer for anchoring randomisation

Value

a dataframe of sensory design with panelist column and order of product presentation columns

Examples

prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  j       d       a       g       k       n       m       b      
#>  2 ind_002  d       g       j       n       a       b       k       e      
#>  3 ind_003  g       n       d       b       j       e       a       f      
#>  4 ind_004  n       b       g       e       d       f       j       h      
#>  5 ind_005  b       e       n       f       g       h       d       l      
#>  6 ind_006  e       f       b       h       n       l       g       i      
#>  7 ind_007  f       h       e       l       b       i       n       c      
#>  8 ind_008  h       l       f       i       e       c       b       m      
#>  9 ind_009  l       i       h       c       f       m       e       k      
#> 10 ind_010  i       c       l       m       h       k       f       a      
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  e (529) f (576) b (300) j (641) g (668) c (347) l (963) k (423)
#>  2 ind_002  f (576) j (641) e (529) c (347) b (300) k (423) g (668) i (275)
#>  3 ind_003  j (641) c (347) f (576) k (423) e (529) i (275) b (300) d (685)
#>  4 ind_004  c (347) k (423) j (641) i (275) f (576) d (685) e (529) m (506)
#>  5 ind_005  k (423) i (275) c (347) d (685) j (641) m (506) f (576) h (515)
#>  6 ind_006  i (275) d (685) k (423) m (506) c (347) h (515) j (641) n (794)
#>  7 ind_007  d (685) m (506) i (275) h (515) k (423) n (794) c (347) a (700)
#>  8 ind_008  m (506) h (515) d (685) n (794) i (275) a (700) k (423) l (963)
#>  9 ind_009  h (515) n (794) m (506) a (700) d (685) l (963) i (275) g (668)
#> 10 ind_010  n (794) a (700) h (515) l (963) m (506) g (668) d (685) b (300)
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # Design of Experiment: <20 x 6>
#> # Panelist:             20 subjects
#> # Product:              5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_04 (689) prod_02 (954) prod_03 (740) prod_01 (658) prod_05 (39…
#>  2 ind_002  prod_02 (954) prod_01 (658) prod_04 (689) prod_05 (396) prod_03 (74…
#>  3 ind_003  prod_01 (658) prod_05 (396) prod_02 (954) prod_03 (740) prod_04 (68…
#>  4 ind_004  prod_05 (396) prod_03 (740) prod_01 (658) prod_04 (689) prod_02 (95…
#>  5 ind_005  prod_03 (740) prod_04 (689) prod_05 (396) prod_02 (954) prod_01 (65…
#>  6 ind_006  prod_05 (396) prod_01 (658) prod_03 (740) prod_02 (954) prod_04 (68…
#>  7 ind_007  prod_03 (740) prod_05 (396) prod_04 (689) prod_01 (658) prod_02 (95…
#>  8 ind_008  prod_04 (689) prod_03 (740) prod_02 (954) prod_05 (396) prod_01 (65…
#>  9 ind_009  prod_02 (954) prod_04 (689) prod_01 (658) prod_03 (740) prod_05 (39…
#> 10 ind_010  prod_01 (658) prod_02 (954) prod_05 (396) prod_04 (689) prod_03 (74…
#> 11 ind_011  prod_04 (689) prod_02 (954) prod_03 (740) prod_01 (658) prod_05 (39…
#> 12 ind_012  prod_02 (954) prod_01 (658) prod_04 (689) prod_05 (396) prod_03 (74…
#> 13 ind_013  prod_01 (658) prod_05 (396) prod_02 (954) prod_03 (740) prod_04 (68…
#> 14 ind_014  prod_05 (396) prod_03 (740) prod_01 (658) prod_04 (689) prod_02 (95…
#> 15 ind_015  prod_03 (740) prod_04 (689) prod_05 (396) prod_02 (954) prod_01 (65…
#> 16 ind_016  prod_05 (396) prod_01 (658) prod_03 (740) prod_02 (954) prod_04 (68…
#> 17 ind_017  prod_03 (740) prod_05 (396) prod_04 (689) prod_01 (658) prod_02 (95…
#> 18 ind_018  prod_04 (689) prod_03 (740) prod_02 (954) prod_05 (396) prod_01 (65…
#> 19 ind_019  prod_02 (954) prod_04 (689) prod_01 (658) prod_03 (740) prod_05 (39…
#> 20 ind_020  prod_01 (658) prod_02 (954) prod_05 (396) prod_04 (689) prod_03 (74…