Specify sensory table

Specify sensory informations into a raw dataframe. The minimal sensory informations are the panelist, the product, the sensory attributes, and the method in which the evaluation was conducted. Additonally the session, the presentation order and hedonic evaluation can also be specified.

specify(data, sensory_method = c("QDA", "CATA", "RATA", "FCP", "FP",
  "JAR", "IPM"), panelist, product, session = NULL, pres_order = NULL,
  attribute, hedonic = NULL)

Arguments

data
a dataframe

sensory_method
method of sensory evaluation, available methods are QDA, CATA, RATA, FCP, FP, JAR, IPM

panelist
panelist column

product
product column

session
session column

pres_order
presentation order column

attribute
sensory attribute columns

hedonic
hedonic column

Value

a sensory table (dataframe with class of tbl_sensory)

Examples

(df <- specify(
  data = perfume_qda_consumers,
  sensory_method = "QDA",
  panelist = consumer,
  product = product,
  attribute = intensity:green,
  hedonic = NULL
))

#> # A sensory table: <1236 x 23>
#> # Sensory method:  Quantitative Descriptive Analysis
#> # Panelist:        consumer <103 subjects>
#> # Product:         product <12 items>
#> # Attribute:       intensity, freshness, jasmin, ro...<21 lexicons>
#> # Hedonic:         None
#>    consumer product intensity freshness jasmin  rose camomille fresh_lemon
#>    <fct>    <fct>       <dbl>     <dbl>  <dbl> <dbl>     <dbl>       <dbl>
#>  1 171      Angel        77        36.1   30    13.3      15.2        15.2
#>  2 171      Aromat…      69.7      26.7   75.2  17.9      15.5        16.7
#>  3 171      Chanel…      52.7      50.6   41.5  45.8      18.8        17.6
#>  4 171      Cinema       70        60.9   26.4  21.8      18.8        20.6
#>  5 171      Coco M…      73        66.1   40.6  18.2      15.2        15.2
#>  6 171      J'ador…      61.8      58.5   30    14.2      16.7        20.9
#>  7 171      J'ador…      48.8      50.9   46.1  48.2      18.2        22.7
#>  8 171      L'inst…      67.3      22.7   15.8  16.1      14.5        36.1
#>  9 171      Lolita…      76.7      19.1   27.9  17        16.1        16.7
#> 10 171      Pleasu…      52.4      67.6   49.4  62.7      20.9        46.4
#> # … with 1,226 more rows, and 15 more variables: vanilla <dbl>, citrus <dbl>,
#> #   anis <dbl>, sweet_fruit <dbl>, honey <dbl>, caramel <dbl>, spicy <dbl>,
#> #   woody <dbl>, leather <dbl>, nutty <dbl>, musk <dbl>, animal <dbl>,
#> #   earthy <dbl>, incense <dbl>, green <dbl>

perfume_qda_experts %>%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = spicy:wrapping
  )

#> # A sensory table: <288 x 16>
#> # Sensory method:  Quantitative Descriptive Analysis
#> # Panelist:        panelist <12 subjects>
#> # Product:         product <12 items>
#> # Attribute:       spicy, heady, fruity, green, van...<12 lexicons>
#> # Hedonic:         None
#>    panelist product session rank  spicy heady fruity green vanilla floral woody
#>    <fct>    <fct>   <fct>   <fct> <dbl> <dbl>  <dbl> <dbl>   <dbl>  <dbl> <dbl>
#>  1 SO       Coco M… 1       1       0.6   0.7    7.1   0.8     2      8.6   0.7
#>  2 SO       Lolita… 1       2       1.4   1.5    3.2   1.3     5.3    4.4   1.1
#>  3 SO       Angel   1       3       3.8   9.7    1     0.6     1.9    3.6   0.6
#>  4 SO       Pure P… 1       4       1.1   1.2    7.4   0.2     2.5    9.5   0.8
#>  5 SO       Chanel… 1       5       4.9   8.4    3     0.3     0.1    4.9   3.3
#>  6 SO       Aromat… 1       6       8.6  10      0     0       3.1    5     1.2
#>  7 SO       J'ador… 1       7       1.5   0      9.4   0.3     0.3    8.1   0.5
#>  8 SO       Pleasu… 1       8       0.2   0.2    5.3   2.8     0.5    9.3   0.5
#>  9 SO       J'ador… 1       9       0.5   0.4    9.6   0.3     0.5    7.5   0.5
#> 10 SO       L'inst… 1       10      2     8      0.5   0.5     4      7     4  
#> # … with 278 more rows, and 5 more variables: citrus <dbl>, marine <dbl>,
#> #   greedy <dbl>, oriental <dbl>, wrapping <dbl>