Prepare sensory experiment

Generate a design of presentation order based on Williams design.

prepare(n_panelist, product, blind_code = FALSE, seed = NULL)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

blind_code
wheteher to generate random three digit number for labeling

seed
an integer for anchoring randomisation

Value

a dataframe of sensory design with panelist column and order of product presentation columns

Examples

prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  g       a       i       c       m       h       n       f      
#>  2 ind_002  a       c       g       h       i       f       m       k      
#>  3 ind_003  c       h       a       f       g       k       i       b      
#>  4 ind_004  h       f       c       k       a       b       g       j      
#>  5 ind_005  f       k       h       b       c       j       a       d      
#>  6 ind_006  k       b       f       j       h       d       c       e      
#>  7 ind_007  b       j       k       d       f       e       h       l      
#>  8 ind_008  j       d       b       e       k       l       f       n      
#>  9 ind_009  d       e       j       l       b       n       k       m      
#> 10 ind_010  e       l       d       n       j       m       b       i      
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  n (532) e (736) b (337) a (431) f (809) h (661) c (667) l (109)
#>  2 ind_002  e (736) a (431) n (532) h (661) b (337) l (109) f (809) j (147)
#>  3 ind_003  a (431) h (661) e (736) l (109) n (532) j (147) b (337) m (632)
#>  4 ind_004  h (661) l (109) a (431) j (147) e (736) m (632) n (532) i (707)
#>  5 ind_005  l (109) j (147) h (661) m (632) a (431) i (707) e (736) g (668)
#>  6 ind_006  j (147) m (632) l (109) i (707) h (661) g (668) a (431) d (155)
#>  7 ind_007  m (632) i (707) j (147) g (668) l (109) d (155) h (661) k (187)
#>  8 ind_008  i (707) g (668) m (632) d (155) j (147) k (187) l (109) c (667)
#>  9 ind_009  g (668) d (155) i (707) k (187) m (632) c (667) j (147) f (809)
#> 10 ind_010  d (155) k (187) g (668) c (667) i (707) f (809) m (632) b (337)
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # Design of Experiment: <20 x 6>
#> # Panelist:             20 subjects
#> # Product:              5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_01 (382) prod_04 (565) prod_03 (707) prod_02 (802) prod_05 (84…
#>  2 ind_002  prod_04 (565) prod_02 (802) prod_01 (382) prod_05 (846) prod_03 (70…
#>  3 ind_003  prod_02 (802) prod_05 (846) prod_04 (565) prod_03 (707) prod_01 (38…
#>  4 ind_004  prod_05 (846) prod_03 (707) prod_02 (802) prod_01 (382) prod_04 (56…
#>  5 ind_005  prod_03 (707) prod_01 (382) prod_05 (846) prod_04 (565) prod_02 (80…
#>  6 ind_006  prod_05 (846) prod_02 (802) prod_03 (707) prod_04 (565) prod_01 (38…
#>  7 ind_007  prod_03 (707) prod_05 (846) prod_01 (382) prod_02 (802) prod_04 (56…
#>  8 ind_008  prod_01 (382) prod_03 (707) prod_04 (565) prod_05 (846) prod_02 (80…
#>  9 ind_009  prod_04 (565) prod_01 (382) prod_02 (802) prod_03 (707) prod_05 (84…
#> 10 ind_010  prod_02 (802) prod_04 (565) prod_05 (846) prod_01 (382) prod_03 (70…
#> 11 ind_011  prod_01 (382) prod_04 (565) prod_03 (707) prod_02 (802) prod_05 (84…
#> 12 ind_012  prod_04 (565) prod_02 (802) prod_01 (382) prod_05 (846) prod_03 (70…
#> 13 ind_013  prod_02 (802) prod_05 (846) prod_04 (565) prod_03 (707) prod_01 (38…
#> 14 ind_014  prod_05 (846) prod_03 (707) prod_02 (802) prod_01 (382) prod_04 (56…
#> 15 ind_015  prod_03 (707) prod_01 (382) prod_05 (846) prod_04 (565) prod_02 (80…
#> 16 ind_016  prod_05 (846) prod_02 (802) prod_03 (707) prod_04 (565) prod_01 (38…
#> 17 ind_017  prod_03 (707) prod_05 (846) prod_01 (382) prod_02 (802) prod_04 (56…
#> 18 ind_018  prod_01 (382) prod_03 (707) prod_04 (565) prod_05 (846) prod_02 (80…
#> 19 ind_019  prod_04 (565) prod_01 (382) prod_02 (802) prod_03 (707) prod_05 (84…
#> 20 ind_020  prod_02 (802) prod_04 (565) prod_05 (846) prod_01 (382) prod_03 (70…