Prepare sensory experiment

Generate a design of presentation order based on Williams design.

prepare(n_panelist, product, blind_code = FALSE)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

blind_code
wheteher to generate random three digit number for labeling

Value

a dataframe of sensory design with panelist column and order of product presentation columns

Examples

prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)

#> # A sensory table: Design of Experiment
#> # Panelist:        30 subjects
#> # Product:         14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  a       h       c       m       e       n       i       k      
#>  2 ind_002  h       m       a       n       c       k       e       g      
#>  3 ind_003  m       n       h       k       a       g       c       d      
#>  4 ind_004  n       k       m       g       h       d       a       f      
#>  5 ind_005  k       g       n       d       m       f       h       l      
#>  6 ind_006  g       d       k       f       n       l       m       b      
#>  7 ind_007  d       f       g       l       k       b       n       j      
#>  8 ind_008  f       l       d       b       g       j       k       i      
#>  9 ind_009  l       b       f       j       d       i       g       e      
#> 10 ind_010  b       j       l       i       f       e       d       c      
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

#> # A sensory table: Design of Experiment
#> # Panelist:        30 subjects
#> # Product:         14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  j (108) c (598) n (705) a (692) f (735) m (978) h (488) b (206)
#>  2 ind_002  c (598) a (692) j (108) m (978) n (705) b (206) f (735) l (581)
#>  3 ind_003  a (692) m (978) c (598) b (206) j (108) l (581) n (705) k (610)
#>  4 ind_004  m (978) b (206) a (692) l (581) c (598) k (610) j (108) i (867)
#>  5 ind_005  b (206) l (581) m (978) k (610) a (692) i (867) c (598) e (283)
#>  6 ind_006  l (581) k (610) b (206) i (867) m (978) e (283) a (692) d (161)
#>  7 ind_007  k (610) i (867) l (581) e (283) b (206) d (161) m (978) g (756)
#>  8 ind_008  i (867) e (283) k (610) d (161) l (581) g (756) b (206) h (488)
#>  9 ind_009  e (283) d (161) i (867) g (756) k (610) h (488) l (581) f (735)
#> 10 ind_010  d (161) g (756) e (283) h (488) i (867) f (735) k (610) n (705)
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # A sensory table: Design of Experiment
#> # Panelist:        20 subjects
#> # Product:         5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_03 (635) prod_02 (576) prod_01 (746) prod_05 (148) prod_04 (54…
#>  2 ind_002  prod_02 (576) prod_05 (148) prod_03 (635) prod_04 (543) prod_01 (74…
#>  3 ind_003  prod_05 (148) prod_04 (543) prod_02 (576) prod_01 (746) prod_03 (63…
#>  4 ind_004  prod_04 (543) prod_01 (746) prod_05 (148) prod_03 (635) prod_02 (57…
#>  5 ind_005  prod_01 (746) prod_03 (635) prod_04 (543) prod_02 (576) prod_05 (14…
#>  6 ind_006  prod_04 (543) prod_05 (148) prod_01 (746) prod_02 (576) prod_03 (63…
#>  7 ind_007  prod_01 (746) prod_04 (543) prod_03 (635) prod_05 (148) prod_02 (57…
#>  8 ind_008  prod_03 (635) prod_01 (746) prod_02 (576) prod_04 (543) prod_05 (14…
#>  9 ind_009  prod_02 (576) prod_03 (635) prod_05 (148) prod_01 (746) prod_04 (54…
#> 10 ind_010  prod_05 (148) prod_02 (576) prod_04 (543) prod_03 (635) prod_01 (74…
#> 11 ind_011  prod_03 (635) prod_02 (576) prod_01 (746) prod_05 (148) prod_04 (54…
#> 12 ind_012  prod_02 (576) prod_05 (148) prod_03 (635) prod_04 (543) prod_01 (74…
#> 13 ind_013  prod_05 (148) prod_04 (543) prod_02 (576) prod_01 (746) prod_03 (63…
#> 14 ind_014  prod_04 (543) prod_01 (746) prod_05 (148) prod_03 (635) prod_02 (57…
#> 15 ind_015  prod_01 (746) prod_03 (635) prod_04 (543) prod_02 (576) prod_05 (14…
#> 16 ind_016  prod_04 (543) prod_05 (148) prod_01 (746) prod_02 (576) prod_03 (63…
#> 17 ind_017  prod_01 (746) prod_04 (543) prod_03 (635) prod_05 (148) prod_02 (57…
#> 18 ind_018  prod_03 (635) prod_01 (746) prod_02 (576) prod_04 (543) prod_05 (14…
#> 19 ind_019  prod_02 (576) prod_03 (635) prod_05 (148) prod_01 (746) prod_04 (54…
#> 20 ind_020  prod_05 (148) prod_02 (576) prod_04 (543) prod_03 (635) prod_01 (74…