Prepare sensory experiment

Generate a design of presentation order based on Williams design.

prepare(n_panelist, product, blind_code = FALSE, seed = NULL)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

blind_code
wheteher to generate random three digit number for labeling

seed
an integer for anchoring randomisation

Value

a dataframe of sensory design with panelist column and order of product presentation columns

Examples

prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  c       a       i       k       b       f       h       e      
#>  2 ind_002  a       k       c       f       i       e       b       n      
#>  3 ind_003  k       f       a       e       c       n       i       m      
#>  4 ind_004  f       e       k       n       a       m       c       g      
#>  5 ind_005  e       n       f       m       k       g       a       l      
#>  6 ind_006  n       m       e       g       f       l       k       j      
#>  7 ind_007  m       g       n       l       e       j       f       d      
#>  8 ind_008  g       l       m       j       n       d       e       h      
#>  9 ind_009  l       j       g       d       m       h       n       b      
#> 10 ind_010  j       d       l       h       g       b       m       i      
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  i (671) h (943) m (323) j (825) b (198) l (553) a (426) e (356)
#>  2 ind_002  h (943) j (825) i (671) l (553) m (323) e (356) b (198) k (967)
#>  3 ind_003  j (825) l (553) h (943) e (356) i (671) k (967) m (323) f (745)
#>  4 ind_004  l (553) e (356) j (825) k (967) h (943) f (745) i (671) d (354)
#>  5 ind_005  e (356) k (967) l (553) f (745) j (825) d (354) h (943) g (812)
#>  6 ind_006  k (967) f (745) e (356) d (354) l (553) g (812) j (825) n (947)
#>  7 ind_007  f (745) d (354) k (967) g (812) e (356) n (947) l (553) c (541)
#>  8 ind_008  d (354) g (812) f (745) n (947) k (967) c (541) e (356) a (426)
#>  9 ind_009  g (812) n (947) d (354) c (541) f (745) a (426) k (967) b (198)
#> 10 ind_010  n (947) c (541) g (812) a (426) d (354) b (198) f (745) m (323)
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # Design of Experiment: <20 x 6>
#> # Panelist:             20 subjects
#> # Product:              5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_02 (198) prod_04 (329) prod_01 (399) prod_03 (156) prod_05 (52…
#>  2 ind_002  prod_04 (329) prod_03 (156) prod_02 (198) prod_05 (521) prod_01 (39…
#>  3 ind_003  prod_03 (156) prod_05 (521) prod_04 (329) prod_01 (399) prod_02 (19…
#>  4 ind_004  prod_05 (521) prod_01 (399) prod_03 (156) prod_02 (198) prod_04 (32…
#>  5 ind_005  prod_01 (399) prod_02 (198) prod_05 (521) prod_04 (329) prod_03 (15…
#>  6 ind_006  prod_05 (521) prod_03 (156) prod_01 (399) prod_04 (329) prod_02 (19…
#>  7 ind_007  prod_01 (399) prod_05 (521) prod_02 (198) prod_03 (156) prod_04 (32…
#>  8 ind_008  prod_02 (198) prod_01 (399) prod_04 (329) prod_05 (521) prod_03 (15…
#>  9 ind_009  prod_04 (329) prod_02 (198) prod_03 (156) prod_01 (399) prod_05 (52…
#> 10 ind_010  prod_03 (156) prod_04 (329) prod_05 (521) prod_02 (198) prod_01 (39…
#> 11 ind_011  prod_02 (198) prod_04 (329) prod_01 (399) prod_03 (156) prod_05 (52…
#> 12 ind_012  prod_04 (329) prod_03 (156) prod_02 (198) prod_05 (521) prod_01 (39…
#> 13 ind_013  prod_03 (156) prod_05 (521) prod_04 (329) prod_01 (399) prod_02 (19…
#> 14 ind_014  prod_05 (521) prod_01 (399) prod_03 (156) prod_02 (198) prod_04 (32…
#> 15 ind_015  prod_01 (399) prod_02 (198) prod_05 (521) prod_04 (329) prod_03 (15…
#> 16 ind_016  prod_05 (521) prod_03 (156) prod_01 (399) prod_04 (329) prod_02 (19…
#> 17 ind_017  prod_01 (399) prod_05 (521) prod_02 (198) prod_03 (156) prod_04 (32…
#> 18 ind_018  prod_02 (198) prod_01 (399) prod_04 (329) prod_05 (521) prod_03 (15…
#> 19 ind_019  prod_04 (329) prod_02 (198) prod_03 (156) prod_01 (399) prod_05 (52…
#> 20 ind_020  prod_03 (156) prod_04 (329) prod_05 (521) prod_02 (198) prod_01 (39…