Prepare sensory experiment

Generate a design of presentation order based on Williams design.

prepare(n_panelist, product, blind_code = FALSE, seed = NULL)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

blind_code
wheteher to generate random three digit number for labeling

seed
an integer for anchoring randomisation

Value

a dataframe of sensory design with panelist column and order of product presentation columns

Examples

prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  n       d       c       b       g       f       j       i      
#>  2 ind_002  d       b       n       f       c       i       g       l      
#>  3 ind_003  b       f       d       i       n       l       c       a      
#>  4 ind_004  f       i       b       l       d       a       n       k      
#>  5 ind_005  i       l       f       a       b       k       d       e      
#>  6 ind_006  l       a       i       k       f       e       b       h      
#>  7 ind_007  a       k       l       e       i       h       f       m      
#>  8 ind_008  k       e       a       h       l       m       i       j      
#>  9 ind_009  e       h       k       m       a       j       l       g      
#> 10 ind_010  h       m       e       j       k       g       a       c      
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  l (409) h (257) d (605) g (648) b (523) n (507) f (150) e (239)
#>  2 ind_002  h (257) g (648) l (409) n (507) d (605) e (239) b (523) c (728)
#>  3 ind_003  g (648) n (507) h (257) e (239) l (409) c (728) d (605) a (371)
#>  4 ind_004  n (507) e (239) g (648) c (728) h (257) a (371) l (409) m (835)
#>  5 ind_005  e (239) c (728) n (507) a (371) g (648) m (835) h (257) i (399)
#>  6 ind_006  c (728) a (371) e (239) m (835) n (507) i (399) g (648) k (918)
#>  7 ind_007  a (371) m (835) c (728) i (399) e (239) k (918) n (507) j (147)
#>  8 ind_008  m (835) i (399) a (371) k (918) c (728) j (147) e (239) f (150)
#>  9 ind_009  i (399) k (918) m (835) j (147) a (371) f (150) c (728) b (523)
#> 10 ind_010  k (918) j (147) i (399) f (150) m (835) b (523) a (371) d (605)
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # Design of Experiment: <20 x 6>
#> # Panelist:             20 subjects
#> # Product:              5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_02 (517) prod_04 (519) prod_01 (555) prod_03 (430) prod_05 (44…
#>  2 ind_002  prod_04 (519) prod_03 (430) prod_02 (517) prod_05 (448) prod_01 (55…
#>  3 ind_003  prod_03 (430) prod_05 (448) prod_04 (519) prod_01 (555) prod_02 (51…
#>  4 ind_004  prod_05 (448) prod_01 (555) prod_03 (430) prod_02 (517) prod_04 (51…
#>  5 ind_005  prod_01 (555) prod_02 (517) prod_05 (448) prod_04 (519) prod_03 (43…
#>  6 ind_006  prod_05 (448) prod_03 (430) prod_01 (555) prod_04 (519) prod_02 (51…
#>  7 ind_007  prod_01 (555) prod_05 (448) prod_02 (517) prod_03 (430) prod_04 (51…
#>  8 ind_008  prod_02 (517) prod_01 (555) prod_04 (519) prod_05 (448) prod_03 (43…
#>  9 ind_009  prod_04 (519) prod_02 (517) prod_03 (430) prod_01 (555) prod_05 (44…
#> 10 ind_010  prod_03 (430) prod_04 (519) prod_05 (448) prod_02 (517) prod_01 (55…
#> 11 ind_011  prod_02 (517) prod_04 (519) prod_01 (555) prod_03 (430) prod_05 (44…
#> 12 ind_012  prod_04 (519) prod_03 (430) prod_02 (517) prod_05 (448) prod_01 (55…
#> 13 ind_013  prod_03 (430) prod_05 (448) prod_04 (519) prod_01 (555) prod_02 (51…
#> 14 ind_014  prod_05 (448) prod_01 (555) prod_03 (430) prod_02 (517) prod_04 (51…
#> 15 ind_015  prod_01 (555) prod_02 (517) prod_05 (448) prod_04 (519) prod_03 (43…
#> 16 ind_016  prod_05 (448) prod_03 (430) prod_01 (555) prod_04 (519) prod_02 (51…
#> 17 ind_017  prod_01 (555) prod_05 (448) prod_02 (517) prod_03 (430) prod_04 (51…
#> 18 ind_018  prod_02 (517) prod_01 (555) prod_04 (519) prod_05 (448) prod_03 (43…
#> 19 ind_019  prod_04 (519) prod_02 (517) prod_03 (430) prod_01 (555) prod_05 (44…
#> 20 ind_020  prod_03 (430) prod_04 (519) prod_05 (448) prod_02 (517) prod_01 (55…