Prepare sensory experiment

Generate a design of presentation order based on Williams design.

prepare(n_panelist, product, blind_code = FALSE, seed = NULL)

Arguments

n_panelist
number of panelist

product
a numeric value of number or product or vector of product names

blind_code
wheteher to generate random three digit number for labeling

seed
an integer for anchoring randomisation

Value

a dataframe of sensory design with panelist column and order of product presentation columns

Examples

prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  l       h       j       g       f       e       d       a      
#>  2 ind_002  h       g       l       e       j       a       f       k      
#>  3 ind_003  g       e       h       a       l       k       j       i      
#>  4 ind_004  e       a       g       k       h       i       l       c      
#>  5 ind_005  a       k       e       i       g       c       h       b      
#>  6 ind_006  k       i       a       c       e       b       g       n      
#>  7 ind_007  i       c       k       b       a       n       e       m      
#>  8 ind_008  c       b       i       n       k       m       a       d      
#>  9 ind_009  b       n       c       m       i       d       k       f      
#> 10 ind_010  n       m       b       d       c       f       i       j      
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

#> # Design of Experiment: <30 x 15>
#> # Panelist:             30 subjects
#> # Product:              14 items
#>    panelist order_1 order_2 order_3 order_4 order_5 order_6 order_7 order_8
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ind_001  j (139) k (433) m (343) g (192) a (912) i (378) n (889) f (298)
#>  2 ind_002  k (433) g (192) j (139) i (378) m (343) f (298) a (912) l (416)
#>  3 ind_003  g (192) i (378) k (433) f (298) j (139) l (416) m (343) h (599)
#>  4 ind_004  i (378) f (298) g (192) l (416) k (433) h (599) j (139) c (561)
#>  5 ind_005  f (298) l (416) i (378) h (599) g (192) c (561) k (433) d (396)
#>  6 ind_006  l (416) h (599) f (298) c (561) i (378) d (396) g (192) b (636)
#>  7 ind_007  h (599) c (561) l (416) d (396) f (298) b (636) i (378) e (517)
#>  8 ind_008  c (561) d (396) h (599) b (636) l (416) e (517) f (298) n (889)
#>  9 ind_009  d (396) b (636) c (561) e (517) h (599) n (889) l (416) a (912)
#> 10 ind_010  b (636) e (517) d (396) n (889) c (561) a (912) h (599) m (343)
#> # … with 20 more rows, and 6 more variables: order_9 <chr>, order_10 <chr>,
#> #   order_11 <chr>, order_12 <chr>, order_13 <chr>, order_14 <chr>

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

#> # Design of Experiment: <20 x 6>
#> # Panelist:             20 subjects
#> # Product:              5 items
#>    panelist order_1       order_2       order_3       order_4       order_5     
#>    <chr>    <chr>         <chr>         <chr>         <chr>         <chr>       
#>  1 ind_001  prod_02 (291) prod_01 (126) prod_03 (625) prod_04 (838) prod_05 (72…
#>  2 ind_002  prod_01 (126) prod_04 (838) prod_02 (291) prod_05 (723) prod_03 (62…
#>  3 ind_003  prod_04 (838) prod_05 (723) prod_01 (126) prod_03 (625) prod_02 (29…
#>  4 ind_004  prod_05 (723) prod_03 (625) prod_04 (838) prod_02 (291) prod_01 (12…
#>  5 ind_005  prod_03 (625) prod_02 (291) prod_05 (723) prod_01 (126) prod_04 (83…
#>  6 ind_006  prod_05 (723) prod_04 (838) prod_03 (625) prod_01 (126) prod_02 (29…
#>  7 ind_007  prod_03 (625) prod_05 (723) prod_02 (291) prod_04 (838) prod_01 (12…
#>  8 ind_008  prod_02 (291) prod_03 (625) prod_01 (126) prod_05 (723) prod_04 (83…
#>  9 ind_009  prod_01 (126) prod_02 (291) prod_04 (838) prod_03 (625) prod_05 (72…
#> 10 ind_010  prod_04 (838) prod_01 (126) prod_05 (723) prod_02 (291) prod_03 (62…
#> 11 ind_011  prod_02 (291) prod_01 (126) prod_03 (625) prod_04 (838) prod_05 (72…
#> 12 ind_012  prod_01 (126) prod_04 (838) prod_02 (291) prod_05 (723) prod_03 (62…
#> 13 ind_013  prod_04 (838) prod_05 (723) prod_01 (126) prod_03 (625) prod_02 (29…
#> 14 ind_014  prod_05 (723) prod_03 (625) prod_04 (838) prod_02 (291) prod_01 (12…
#> 15 ind_015  prod_03 (625) prod_02 (291) prod_05 (723) prod_01 (126) prod_04 (83…
#> 16 ind_016  prod_05 (723) prod_04 (838) prod_03 (625) prod_01 (126) prod_02 (29…
#> 17 ind_017  prod_03 (625) prod_05 (723) prod_02 (291) prod_04 (838) prod_01 (12…
#> 18 ind_018  prod_02 (291) prod_03 (625) prod_01 (126) prod_05 (723) prod_04 (83…
#> 19 ind_019  prod_01 (126) prod_02 (291) prod_04 (838) prod_03 (625) prod_05 (72…
#> 20 ind_020  prod_04 (838) prod_01 (126) prod_05 (723) prod_02 (291) prod_03 (62…